Fix install, add mechanism rejecting outdated
During install, we were resolving the `install` promise on `close` of
the tar writer or on `close` of the http download. Which is obviously
wrong - we must wait for tar writer only. Otherwise we could resolve
before the tar contents have been fully written, which is exactly what
was happening: Cannot start gel-ls: spawn ETXTBSY

---

I've also added a machanism to parse the version of gel-ls and reject
it if it is too old.

The idea here is that when we add a new non-backward-compatible feature
to vscode extension (such as command line options), we bump
`minGelVersion` and force users to update to a recent version of gel-ls.